[fe] more refactoring
