+++ web-client/src/admin/UserList.tsx
@@ -1,6 +1,7 @@
 import { h } from 'preact';
-import Container from '../components/Container';
-import Header from '../components/Header';
+
+import Container from '~/components/Container';
+import Header from '~/components/Header';
 
 function UserList(): h.JSX.Element {
     return (
