+++ web-client/src/auth/AuthContext.tsx
@@ -1,6 +1,7 @@
-import { Session } from './SignInForm';
 import { createContext, h } from 'preact';
 import { route } from 'preact-router';
+
+import { Session } from '~/models/Session';
 
 const TOKEN_KEY = 'access_token';
 const EXPIRATION_KEY = 'access_expiration';
