+++ web-client/src/components/Container.tsx
@@ -1,10 +1,12 @@
-import '../style/Dashboard.scss';
-
 import { h } from 'preact';
-import { SITE_NAME } from '../config';
 import { useContext } from 'preact/hooks';
-import Auth from '../auth/AuthContext';
+
+import { SITE_NAME } from '~/config';
+import Auth from '~/auth/AuthContext';
+
 import Sidebar from './Sidebar';
+
+import '../style/Dashboard.scss';
 
 interface Props {
     children: h.JSX.Element | Array<h.JSX.Element>;
