+++ web-client/src/models/Resource.ts
