+++ web-client/src/models/Session.ts
@@ -1,0 +1,4 @@
+export interface Session {
+    token: string;
+    expires_at: string;
+}
