+++ web-client/src/models/User.ts
@@ -1,0 +1,4 @@
+export interface User {
+    email_address: string;
+    full_name: string;
+}
