+++ web-client/src/tasks/NewTaskForm.tsx
@@ -1,6 +1,8 @@
 import { h } from 'preact';
 import { useContext, useState } from 'preact/hooks';
-import Auth from '../auth/AuthContext';
+
+import Auth from '~/auth/AuthContext';
+
 import createTask from './createTask';
 
 export default function NewTaskForm(): h.JSX.Element {
